update: fix bug

+++ lib/data/hive_data_store.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 
-///
 import '../model/task.model.dart';
 
 class HiveDataStore {
@@ -9,8 +8,11 @@ class HiveDataStore {
   final Box<Task> box = Hive.box<Task>(boxName);
 
   /// Add new Task
+  //coverage:ignore-start
   Future<void> addTask({required Task task}) async {
-    await box.put(task.id, task);
+    if (task.id!.isNotEmpty && task.id != null) {
+      await box.put(task.id, task);
+    }
   }
 
   /// Show task
@@ -25,10 +27,13 @@ class HiveDataStore {
 
   /// Delete task
   Future<void> deleteTask({required Task task}) async {
-    await task.delete();
+    if (task.id!.isNotEmpty && task.id != null) {
+      await task.delete();
+    }
   }
 
   ValueListenable<Box<Task>> listenToTask() {
     return box.listenable();
+    //coverage:ignore-end
   }
 }
